split event properties into own scripts
