+++ Assets/Scenes/Editor.unity
@@ -4799,20 +4799,13 @@ MonoBehaviour:
   m_GameObject: {fileID: 184838087}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3b7c76a246115c1459c963e93f7db056, type: 3}
+  m_Script: {fileID: 11500000, guid: 8ada001011e54c74b87c04d7186d5f3c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   caption: {fileID: 44496736}
-  parameterManager: {fileID: 830452354}
-  slider: {fileID: 0}
-  inputField: {fileID: 0}
-  toggle: {fileID: 0}
+  parameterManager: {fileID: 0}
+  propertyName: 
   dropdown: {fileID: 9172948791891367920}
-  ColorBTN: {fileID: 0}
-  ColorTable: {fileID: 0}
-  colorTableActive: 0
-  colorPreview: {fileID: 0}
-  inputFieldString: {fileID: 0}
 --- !u!1 &189878587
 GameObject:
   m_ObjectHideFlags: 0
@@ -7942,20 +7935,14 @@ MonoBehaviour:
   m_GameObject: {fileID: 345301789}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3b7c76a246115c1459c963e93f7db056, type: 3}
+  m_Script: {fileID: 11500000, guid: 6e9ad350a96f5644dbb1e4686a6bcaed, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   caption: {fileID: 1882704219}
-  parameterManager: {fileID: 830452354}
+  parameterManager: {fileID: 0}
+  propertyName: 
   slider: {fileID: 749529046}
   inputField: {fileID: 735918245}
-  toggle: {fileID: 0}
-  dropdown: {fileID: 0}
-  ColorBTN: {fileID: 0}
-  ColorTable: {fileID: 0}
-  colorTableActive: 0
-  colorPreview: {fileID: 0}
-  inputFieldString: {fileID: 0}
 --- !u!224 &345301791
 RectTransform:
   m_ObjectHideFlags: 0
@@ -8251,15 +8238,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   caption: {fileID: 419538178}
   parameterManager: {fileID: 830452354}
-  slider: {fileID: 0}
-  inputField: {fileID: 606981307}
-  toggle: {fileID: 0}
-  dropdown: {fileID: 0}
-  ColorBTN: {fileID: 0}
-  ColorTable: {fileID: 0}
-  colorTableActive: 0
-  colorPreview: {fileID: 0}
-  inputFieldString: {fileID: 0}
+  propertyName: 
 --- !u!1 &365690205
 GameObject:
   m_ObjectHideFlags: 0
@@ -8604,20 +8583,14 @@ MonoBehaviour:
   m_GameObject: {fileID: 396879232}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3b7c76a246115c1459c963e93f7db056, type: 3}
+  m_Script: {fileID: 11500000, guid: 6e9ad350a96f5644dbb1e4686a6bcaed, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   caption: {fileID: 5490985}
-  parameterManager: {fileID: 830452354}
+  parameterManager: {fileID: 0}
+  propertyName: 
   slider: {fileID: 2067224779}
   inputField: {fileID: 1182082595}
-  toggle: {fileID: 0}
-  dropdown: {fileID: 0}
-  ColorBTN: {fileID: 0}
-  ColorTable: {fileID: 0}
-  colorTableActive: 0
-  colorPreview: {fileID: 0}
-  inputFieldString: {fileID: 0}
 --- !u!1 &407289014
 GameObject:
   m_ObjectHideFlags: 0
@@ -9586,15 +9559,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   caption: {fileID: 0}
   parameterManager: {fileID: 830452354}
-  slider: {fileID: 390818399}
-  inputField: {fileID: 2031558985}
-  toggle: {fileID: 0}
-  dropdown: {fileID: 0}
-  ColorBTN: {fileID: 0}
-  ColorTable: {fileID: 0}
-  colorTableActive: 0
-  colorPreview: {fileID: 0}
-  inputFieldString: {fileID: 0}
+  propertyName: 
 --- !u!1 &435201908
 GameObject:
   m_ObjectHideFlags: 0
@@ -17147,20 +17112,13 @@ MonoBehaviour:
   m_GameObject: {fileID: 755586977}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3b7c76a246115c1459c963e93f7db056, type: 3}
+  m_Script: {fileID: 11500000, guid: 9aa690f14ccbf9e4bb6bd339d500c3e7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   caption: {fileID: 774969486}
-  parameterManager: {fileID: 830452354}
-  slider: {fileID: 0}
-  inputField: {fileID: 0}
+  parameterManager: {fileID: 0}
+  propertyName: 
   toggle: {fileID: 279538359}
-  dropdown: {fileID: 0}
-  ColorBTN: {fileID: 0}
-  ColorTable: {fileID: 0}
-  colorTableActive: 0
-  colorPreview: {fileID: 0}
-  inputFieldString: {fileID: 0}
 --- !u!1 &767779042
 GameObject:
   m_ObjectHideFlags: 0
@@ -31745,20 +31703,16 @@ MonoBehaviour:
   m_GameObject: {fileID: 1443721746}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3b7c76a246115c1459c963e93f7db056, type: 3}
+  m_Script: {fileID: 11500000, guid: c1c576a0586b70d4395de537078023d5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   caption: {fileID: 537615369}
-  parameterManager: {fileID: 830452354}
-  slider: {fileID: 0}
-  inputField: {fileID: 0}
-  toggle: {fileID: 0}
-  dropdown: {fileID: 0}
+  parameterManager: {fileID: 0}
+  propertyName: 
   ColorBTN: {fileID: 1535224918}
   ColorTable: {fileID: 723094753}
   colorTableActive: 0
   colorPreview: {fileID: 1535224921}
-  inputFieldString: {fileID: 0}
 --- !u!1 &1453521077
 GameObject:
   m_ObjectHideFlags: 0
@@ -36133,19 +36087,12 @@ MonoBehaviour:
   m_GameObject: {fileID: 1671827686}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3b7c76a246115c1459c963e93f7db056, type: 3}
+  m_Script: {fileID: 11500000, guid: c01fcc0bb14adee46a4869c1c009850e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   caption: {fileID: 1478022678}
-  parameterManager: {fileID: 830452354}
-  slider: {fileID: 0}
-  inputField: {fileID: 0}
-  toggle: {fileID: 0}
-  dropdown: {fileID: 0}
-  ColorBTN: {fileID: 0}
-  ColorTable: {fileID: 0}
-  colorTableActive: 0
-  colorPreview: {fileID: 0}
+  parameterManager: {fileID: 0}
+  propertyName: 
   inputFieldString: {fileID: 1776034728}
 --- !u!1 &1682022763
 GameObject:
