+++ Assets/Scripts/LevelEditor/EventSelector/PropertyPrefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b86ddd3d47bb04e48a61db47242e02ed
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
